feat(windows/python_whl): make windows HAPPY for build megbrain python package
GitOrigin-RevId: 92b2c07bf939efcac6432bf4c93cd5c46910f129

+++ src/serialization/include/megbrain/serialization/opr_registry.h
@@ -76,8 +76,7 @@ namespace serialization {
 
 #if MGB_ENABLE_DEBUG_UTIL
         //! dump registered oprs
-        static std::vector<std::pair<unsigned long int, std::string>>
-        dump_registries();
+        static std::vector<std::pair<size_t, std::string>> dump_registries();
 #endif
     };
